Added 'force_nonsig' argument to function.
This argument, when set to F, allows studies to reach statistical significance as power and sampling variation permit. Thus, toggling this to F makes percent_sig a rough lower bound, rather than the expected value of the proportion of significant results.

The old behavior is preseved under the default force_nonsig = T. This forces exactly the proportion of significant results. If power > percent_sig, then, this will lead to a downward bias in effect size estimates.