fixed another bug in resampling routine
when resamples were taken using sample_n(), the number of studies resampled was previously set to length(pc_data). length(pc_data) will always return the number of columns, and NOT the number of rows (which is what we want). This had the effect of producing outrageously large CIs. Now using length(pc_data$t_obs) to get resamples with the same number of rows.